chore: remove unused paging module
Paging module contains the old way of implementing tables with sorting/filtering/paging that is no longer used (See #247).  It was kept around to give downstream projects time to migrate.
