--- src/app/common/paging/pageable-table/pageable-table.component.html
@@ -1,117 +1,0 @@
-<div class="table" [ngClass]="{ 'card p-3': showInCard }">
-	<div class="table-header align-items-center mb-3" *ngIf="pagerAtTop || showActions">
-		<div
-			class="table-actions d-flex flex-grow-1"
-			*ngIf="showActions"
-			[hidden]="itemsInternal.length === 0 && hideActionsNoData"
-		>
-			<ng-container *ngTemplateOutlet="actionTemplate"></ng-container>
-		</div>
-
-		<div
-			class="table-header-pager"
-			*ngIf="pagerAtTop"
-			[ngClass]="{ 'ml-auto': !showActions }"
-			[hidden]="itemsInternal.length === 0"
-		>
-			<asy-pager
-				[pageNumber]="pagingOptions.pageNumber"
-				[pageSize]="pagingOptions.pageSize"
-				[totalSize]="pagingOptions.totalSize"
-				[maxPageSize]="pagingOptions.pageSize"
-				[currentSize]="itemsInternal.length"
-				[disableGoToEnd]="disableGoToEnd"
-				(pageChange)="pageChange.emit($event)"
-			>
-			</asy-pager>
-		</div>
-	</div>
-
-	<div
-		class="table-content table-responsive"
-		[hidden]="hideTableNoData && itemsInternal.length === 0"
-	>
-		<!-- Actual Results -->
-		<table
-			class="table"
-			[ngClass]="{ 'table-striped': tableStriped, 'table-hover': tableHover }"
-			aria-describedby="Results table"
-		>
-			<thead>
-				<tr>
-					<th scope="col" *ngIf="rowExpandedTemplate" class="table-action-expand"></th>
-					<ng-container *ngTemplateOutlet="headerTemplate"></ng-container>
-				</tr>
-			</thead>
-
-			<tbody>
-				<ng-container *ngFor="let item of itemsInternal; let i = index">
-					<tr class="table-row">
-						<td
-							*ngIf="rowExpandedTemplate"
-							(click)="itemsExpanded[i] = !itemsExpanded[i]"
-							class="table-action-expand pointer"
-						>
-							<span
-								class="fa fa-2x {{
-									itemsExpanded[i] ? 'fa-angle-up' : 'fa-angle-down'
-								}}"
-							></span>
-						</td>
-						<ng-container
-							*ngTemplateOutlet="rowTemplate; context: { item: item, index: i }"
-						></ng-container>
-					</tr>
-					<ng-container *ngIf="rowExpandedTemplate && itemsExpanded[i]">
-						<tr *ngIf="tableStriped" class="d-none"></tr>
-						<!-- Dummy tr to keep row color the same when table-striped -->
-						<tr class="table-row table-row-expanded">
-							<ng-container
-								*ngTemplateOutlet="
-									rowExpandedTemplate;
-									context: { item: item, index: i }
-								"
-							></ng-container>
-						</tr>
-					</ng-container>
-				</ng-container>
-			</tbody>
-		</table>
-	</div>
-
-	<div class="table-no-results" *ngIf="!loading && hasItems && itemsInternal.length === 0">
-		<ng-container *ngTemplateOutlet="noResultsTableTemplate"></ng-container>
-	</div>
-
-	<div class="table-no-data" *ngIf="!hasItems && !loading">
-		<ng-container *ngTemplateOutlet="noDataTableTemplate"></ng-container>
-	</div>
-
-	<div class="table-footer align-items-center" *ngIf="pagerAtBottom || showFooterActions">
-		<div
-			class="table-actions d-flex flex-grow-1"
-			*ngIf="showFooterActions"
-			[hidden]="itemsInternal.length === 0 && hideFooterActionsNoData"
-		>
-			<ng-container *ngTemplateOutlet="footerActionTemplate"></ng-container>
-		</div>
-
-		<div
-			class="table-footer-pager"
-			*ngIf="pagerAtBottom"
-			[ngClass]="{ 'ml-auto': !showFooterActions }"
-			[hidden]="itemsInternal.length === 0"
-		>
-			<asy-pager
-				[pageNumber]="pagingOptions.pageNumber"
-				[pageSize]="pagingOptions.pageSize"
-				[totalSize]="pagingOptions.totalSize"
-				[maxPageSize]="pagingOptions.pageSize"
-				[currentSize]="itemsInternal.length"
-				[disableGoToEnd]="disableGoToEnd"
-				(pageChange)="pageChange.emit($event)"
-			>
-			</asy-pager>
-		</div>
-	</div>
-</div>
