--- src/app/common/paging/pageable-table/pageable-table.component.scss
@@ -1,12 +1,0 @@
-.table-header,
-.table-footer {
-	display: flex;
-}
-
-.table-row-expanded ::ng-deep td {
-	border-top: none;
-}
-
-[hidden] {
-	display: none !important;
-}
