--- src/app/common/paging/pager/pager.component.html
@@ -1,62 +1,0 @@
-<!-- Paging Controls -->
-<div
-	*ngIf="
-		totalSize > 0 ||
-			((totalSize ?? null) === null && currentSize > 0) ||
-			(pageNumber ?? 0 !== 0);
-		else invalidPage
-	"
->
-	<span class="pager-controls">
-		<!-- Count message -->
-		<span class="pager-status">
-			{{ startFormatted }} - {{ endFormatted }} of {{ totalFormatted }}
-			<span
-				class="warning"
-				class="fa fa-warning"
-				*ngIf="showCountWarning"
-				tooltip="{{ countWarningMessage }}"
-			></span>
-		</span>
-
-		<!-- Paging controls -->
-		<button
-			type="button"
-			class="btn btn-link"
-			[hidden]="totalPages <= 1 || pageNumber - 1 < 0"
-			(click)="goToPage(0)"
-		>
-			<span class="fa fa-angle-double-left"></span>
-		</button>
-
-		<button
-			type="button"
-			class="btn btn-link"
-			[hidden]="totalPages <= 1 || pageNumber - 1 < 0"
-			(click)="goToPage(pageNumber - 1)"
-		>
-			<span class="fa fa-angle-left"></span>
-		</button>
-
-		<button
-			type="button"
-			class="btn btn-link"
-			[hidden]="totalPages <= 1 || pageNumber + 1 >= totalPages || currentSize !== pageSize"
-			(click)="goToPage(pageNumber + 1)"
-		>
-			<span class="fa fa-angle-right"></span>
-		</button>
-
-		<button
-			type="button"
-			class="btn btn-link"
-			[hidden]="totalPages <= 1 || pageNumber + 1 >= totalPages || currentSize !== pageSize"
-			(click)="goToPage(totalPages - 1)"
-			*ngIf="!disableGoToEnd"
-		>
-			<span class="fa fa-angle-double-right"></span>
-		</button>
-	</span>
-</div>
-
-<ng-template #invalidPage>- of -</ng-template>
