--- src/app/common/paging/quick-filters/quick-filters.component.scss
@@ -1,29 +1,0 @@
-@import '../../../../styles/shared';
-
-.quick-select {
-	margin-bottom: 1rem;
-	padding: 15px 15px 10px;
-
-	.collapsible {
-		cursor: pointer;
-	}
-
-	.quick-select-title {
-		@include font-bold(16px);
-		line-height: 16px;
-	}
-
-	label {
-		margin-bottom: 0;
-	}
-
-	input[type='checkbox'] {
-		margin-right: 5px;
-
-		&:checked {
-			+ label {
-				font-family: $font-family-bold;
-			}
-		}
-	}
-}
