--- src/app/common/paging/sort-controls/sort-controls.component.html
@@ -1,21 +1,0 @@
-<div class="paging-sort">
-	<span *ngIf="showHeader">Sort By:</span>
-	<span *ngFor="let option of sortOptions | keys; let i = index">
-		<span *ngIf="i > 0"> / </span>
-		<button
-			type="button"
-			class="btn btn-sm btn-link"
-			[class.selected]="selectedKey === option.value.sortField"
-			(click)="updateSort(option.key)"
-		>
-			{{ option.value.label }}
-			<span
-				*ngIf="showDirectionControl && selectedKey === option.value.sortField"
-				style="display: inline-block"
-			>
-				<span *ngIf="selectedDir === 'DESC'" class="fa fa-sort-amount-desc"></span>
-				<span *ngIf="selectedDir === 'ASC'" class="fa fa-sort-amount-asc"></span>
-			</span>
-		</button>
-	</span>
-</div>
