--- src/app/common/paging/sortable-table-header/sortable-table-header.component.scss
@@ -1,22 +1,0 @@
-@import '../../../../styles/shared';
-
-.low-opacity {
-	opacity: 0.3;
-}
-
-.sorted {
-	color: $ux-color-primary;
-	opacity: 1;
-}
-
-.fa-stack {
-	font-size: 1rem;
-	height: 6px;
-	line-height: 6px;
-	width: 22px;
-}
-
-.sortable-table-header.sortable {
-	cursor: pointer;
-	padding-right: 5px;
-}
