+++ src/app/common/sorting.model.ts
@@ -1,0 +1,11 @@
+export class SortDirection {
+	public static readonly desc = 'DESC';
+	public static readonly asc = 'ASC';
+}
+
+export type SortDir = 'DESC' | 'ASC';
+
+export interface SortChange {
+	sortField: string | string[];
+	sortDir: SortDir;
+}
