+++ src/app/core/admin/feedback/list-feedback/admin-list-feedback.component.spec.ts
@@ -8,7 +8,7 @@ import { NgSelectModule } from '@ng-select/ng-select';
 import { TooltipModule } from 'ngx-bootstrap/tooltip';
 import { of } from 'rxjs';
 
-import { PagingResults } from '../../../../common/paging/paging.model';
+import { PagingResults } from '../../../../common/paging.model';
 import { PipesModule } from '../../../../common/pipes.module';
 import { SearchInputModule } from '../../../../common/search-input.module';
 import { SystemAlertModule } from '../../../../common/system-alert.module';
