+++ src/app/core/audit/audit.service.ts
@@ -4,7 +4,7 @@ import { Injectable } from '@angular/core';
 import { of, Observable } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
 
-import { NULL_PAGING_RESULTS, PagingOptions, PagingResults } from '../../common/paging.module';
+import { NULL_PAGING_RESULTS, PagingOptions, PagingResults } from '../../common/paging.model';
 import { SystemAlertService } from '../../common/system-alert/system-alert.service';
 import { User } from '../auth/user.model';
 import { AuditActionTypes } from './audit.classes';
