+++ src/app/core/audit/list-audit-entries/audit-actor-filter.directive.ts
@@ -3,7 +3,7 @@ import { Directive, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
-import { PagingOptions } from '../../../common/paging/paging.model';
+import { PagingOptions } from '../../../common/paging.model';
 import { AsyHeaderTypeaheadFilterComponent } from '../../../common/table/filter/asy-header-typeahead-filter/asy-header-typeahead-filter.component';
 import { AuditService } from '../audit.service';
