+++ src/app/core/teams/list-team-members/list-team-members.component.ts
@@ -17,8 +17,9 @@ import { catchError, filter, first, switchMap } from 'rxjs/operators';
 
 import { ModalAction } from '../../../common/modal/modal.model';
 import { ModalService } from '../../../common/modal/modal.service';
-import { PagingOptions, PagingResults, SortDirection } from '../../../common/paging.module';
+import { PagingOptions, PagingResults } from '../../../common/paging.model';
 import { isNotNullOrUndefined } from '../../../common/rxjs-utils';
+import { SortDirection } from '../../../common/sorting.model';
 import { SystemAlertService } from '../../../common/system-alert/system-alert.service';
 import { AsyTableDataSource } from '../../../common/table/asy-table-data-source';
 import { AsyFilterDirective } from '../../../common/table/filter/asy-filter.directive';
